// src/frontend/src/declarations/backend.did.d.ts
/* eslint-disable */

// @ts-nocheck

// This file was automatically generated by @caffeinelabs/bindgen@0.1.0.
// You should NOT make any changes in this file as it will be overwritten.
// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.

import type { ActorMethod } from '@icp-sdk/core/agent';
import type { IDL } from '@icp-sdk/core/candid';
import type { Principal } from '@icp-sdk/core/principal';

export interface MembershipPlan {
  'startTime' : Ms,
  'endTime' : Ms,
  'planType' : PlanType,
}
export type Ms = bigint;
export type PlanType = { 'annual' : null } |
  { 'monthly' : null };
export interface TrialInfo { 'startTime' : Ms, 'duration' : Ms }
export interface UserAccessStatus {
  'canAccess' : boolean,
  'hasActiveTrial' : boolean,
  'hasActiveMembership' : boolean,
}
export type UserId = Principal;
export interface UserProfile { 'name' : string }
export type UserRole = { 'admin' : null } |
  { 'user' : null } |
  { 'guest' : null };
export interface _CaffeineStorageCreateCertificateResult {
  'method' : string,
  'blob_hash' : string,
}
export interface _CaffeineStorageRefillInformation {
  'proposed_top_up_amount' : [] | [bigint],
}
export interface _CaffeineStorageRefillResult {
  'success' : [] | [boolean],
  'topped_up_amount' : [] | [bigint],
}
export interface _SERVICE {
  '_caffeineStorageBlobIsLive' : ActorMethod<[Uint8Array], boolean>,
  '_caffeineStorageBlobsToDelete' : ActorMethod<[], Array<Uint8Array>>,
  '_caffeineStorageConfirmBlobDeletion' : ActorMethod<
    [Array<Uint8Array>],
    undefined
  >,
  '_caffeineStorageCreateCertificate' : ActorMethod<
    [string],
    _CaffeineStorageCreateCertificateResult
  >,
  '_caffeineStorageRefillCashier' : ActorMethod<
    [[] | [_CaffeineStorageRefillInformation]],
    _CaffeineStorageRefillResult
  >,
  '_caffeineStorageUpdateGatewayPrincipals' : ActorMethod<[], undefined>,
  '_initializeAccessControlWithSecret' : ActorMethod<[string], undefined>,
  'activateMembership' : ActorMethod<[UserId, PlanType], MembershipPlan>,
  'assignCallerUserRole' : ActorMethod<[Principal, UserRole], undefined>,
  'getAllMemberships' : ActorMethod<[], Array<[UserId, MembershipPlan]>>,
  'getCallerUserProfile' : ActorMethod<[], [] | [UserProfile]>,
  'getCallerUserRole' : ActorMethod<[], UserRole>,
  'getTrialInfo' : ActorMethod<[], [] | [TrialInfo]>,
  'getUserAccessStatus' : ActorMethod<[], UserAccessStatus>,
  'getUserProfile' : ActorMethod<[Principal], [] | [UserProfile]>,
  'isCallerAdmin' : ActorMethod<[], boolean>,
  'saveCallerUserProfile' : ActorMethod<[UserProfile], undefined>,
  'startTrial' : ActorMethod<[], TrialInfo>,
}
export declare const idlService: IDL.ServiceClass;
export declare const idlInitArgs: IDL.Type[];
export declare const idlFactory: IDL.InterfaceFactory;
export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];